move auto-detect to workerFactory
