--- src/util/default_worker_url.ts
@@ -1,11 +1,0 @@
-export function defaultWorkerUrl(): string {
-    try {
-        const moduleUrl = import.meta.url;
-        const workerName = moduleUrl.endsWith('-dev.mjs')
-            ? 'maplibre-gl-worker-dev.mjs'
-            : 'maplibre-gl-worker.mjs';
-        return new URL(`./${workerName}`, moduleUrl).href;
-    } catch {
-        return '';
-    }
-}
